chore: restructure dist layout to work with existing tests and examples
