+++ examples/floor.mts
@@ -1,4 +1,4 @@
-import { Decimal128 } from "../src/decimal128.mjs";
+import { Decimal128 } from "../dist/esm/decimal128.mjs";
 
 function floor(d: Decimal128): Decimal128 {
     return d.round(0, "floor");
