+++ examples/round.js
@@ -2,7 +2,7 @@
 
 // Examples from the Intl.NumberFormat spec
 
-import { Decimal128 } from "../src/decimal128.mjs";
+import { Decimal128 } from "../dist/esm/decimal128.mjs";
 
 let minusOnePointFive = new Decimal128("-1.5");
 let zeroPointFour = new Decimal128("0.4");
