+++ tests/round.test.js
@@ -1,4 +1,4 @@
-import { Decimal128 } from "../src/decimal128.mjs";
+import { Decimal128 } from "../dist/esm/decimal128.mjs";
 import * as string_decoder from "string_decoder";
 import { expectDecimal128 } from "./util.js";
